% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figlet.R
\name{figlet_list_fonts}
\alias{figlet_list_fonts}
\title{List available fonts in FIGlet.jl}
\usage{
figlet_list_fonts()
}
\value{
Character vector of font names.
}
\description{
List available fonts in FIGlet.jl
}
\examples{
if (interactive()) {
  figlet_list_fonts()
}
}
