% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rjfiglet-package.R
\docType{package}
\name{rjfiglet-package}
\alias{rjfiglet}
\alias{rjfiglet-package}
\title{rjfiglet: R Interface to Julia's FIGlet.jl for ASCII Art}
\description{
Provides an R interface to the Julia package FIGlet.jl for generating ASCII art text banners. This package uses juliaCall to call Julia functions from R, allowing R users to access figlet functionality.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/yourusername/rjfiglet}
  \item Report bugs at \url{https://github.com/yourusername/rjfiglet/issues}
}

}
\author{
\strong{Maintainer}: Guangchuang Yu \email{guangchuangyu@gmail.com}

}
\keyword{internal}
